requirements.txt
Flask==2.0.3
confluent-kafka==2.10.0
etcd3==0.12.0
protobuf==3.18.1
requests==2.28.2
docker==5.0.3
PyYAML==6.0
python-memcached==1.59
psutil==5.9.0
ttkbootstrap==1.10.1
prettytable==3.6.0
configparser==5.3.0
readerwriterlock==1.0.9
